renamed all-supported-tokens.json to main.json
