+++ manifests/main.json
