+++ tools/coingecko/get-tokens-only-with-coingecko-source.js
@@ -1,5 +1,5 @@
 const allSupportedConfig =
-  require("../../manifests/all-supported-tokens.json");
+  require("../../manifests/main.json");
 
 main();
